refactor: proper includes and avoid mess
