+++ qemu_share/fabricmanager/run_server.sh
