+++ qemu_share/includes/ioctl_defs.h
@@ -1,0 +1,10 @@
+#ifndef CXL_SWITCH_IOCTL_DEFS
+#define CXL_SWITCH_IOCTL_DEFS
+
+#include <asm-generic/ioctl.h>
+
+#define CXL_SWITCH_IOCTL_MAGIC 'c'
+#define CXL_SWITCH_IOCTL_SET_EVENTFD_NOTIFY    _IOW(CXL_SWITCH_IOCTL_MAGIC, 1, int)
+#define CXL_SWITCH_IOCTL_SET_EVENTFD_CMD_READY _IOW(CXL_SWITCH_IOCTL_MAGIC, 2, int)
+
+#endif
