+++ qemu_share/server/build_server.sh
