+++ qemu_share/server/diancie_main.cpp
@@ -1,4 +1,4 @@
-#include "../cxl_switch_ipc.h"
+#include "../includes/cxl_switch_ipc.h"
 #include "rpcserver.hpp"
 #include <exception>
 #include <ios>
