+++ qemu_share/server/rpc/rpcclient.hpp
@@ -1,0 +1,8 @@
+#ifndef DIANCIE_RPC_CLIENT_HPP
+#define DIANCIE_RPC_CLIENT_HPP
+
+#include "../cxl_switch_ipc.h"
+
+
+
+#endif
